feat(pokemon): refactor and implement view transition
