+++ src/app/pokemon/detail/[name]/components/Content/View.module.css
