+++ src/app/pokemon/detail/[name]/components/Content/components/Evolution/View.types.ts
