+++ src/app/pokemon/detail/[name]/components/Content/components/Moves/View.module.css
