+++ src/app/pokemon/detail/[name]/components/Content/components/Moves/index.ts
