+++ src/app/pokemon/detail/[name]/components/Content/components/Stats/View.tsx
