+++ src/app/pokemon/detail/[name]/components/Content/components/Stats/View.types.ts
@@ -1,4 +1,4 @@
-import { PokemonData } from "../../View.types";
+import { PokemonData } from "@/app/pokemon/types";
 
 export interface StatsProps {
   stats: NonNullable<PokemonData["stats"]>;
