+++ src/app/pokemon/detail/[name]/components/Content/index.ts
