+++ src/app/pokemon/detail/[name]/components/MainImage/components/ImageWrapper/View.module.css
