+++ src/app/pokemon/detail/[name]/components/MainImage/components/ImageWrapper/View.types.ts
@@ -1,0 +1,7 @@
+import { ReactNode } from "react";
+
+export interface ImageWrapperProps {
+  children: ReactNode;
+  id: number;
+  name: string;
+}
