+++ src/app/pokemon/detail/[name]/components/MainImage/components/ImageWrapper/components/CatchDialog/View.tsx
