+++ src/app/pokemon/detail/[name]/components/MainImage/components/ImageWrapper/components/CatchDialog/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./View";
