+++ src/app/pokemon/detail/[name]/components/MainImage/components/ImageWrapper/components/FloatingBall/View.tsx
