+++ src/app/pokemon/detail/[name]/components/MainImage/components/ImageWrapper/index.ts
