+++ src/app/pokemon/types/index.ts
@@ -7,3 +7,29 @@ export interface PokeLS {
 }
 
 export type EditMode = "" | "edit" | "release";
+
+interface MovesData {
+  move?: {
+    name?: string;
+  };
+}
+
+interface StatsData {
+  base_stat?: number;
+  stat?: {
+    name?: string;
+  };
+}
+
+interface TypesData {
+  type?: {
+    name?: string;
+  };
+}
+
+export interface PokemonData {
+  id?: number;
+  moves?: MovesData[];
+  stats?: StatsData[];
+  types?: TypesData[];
+}
