+++ src/utils/getLS/index.ts
@@ -1,6 +1,11 @@
+import canUseDOM from "../canUseDOM";
+
 const getLS = <T>(name: string): T | null => {
-  const result = localStorage.getItem(name);
-  return result ? JSON.parse(result) : result;
+  if (canUseDOM) {
+    const result = localStorage.getItem(name);
+    return result ? JSON.parse(result) : (result as T | null);
+  }
+  return null;
 };
 
 export default getLS;
